requirements.txt
langchain
langchain-core

langchain-openai
openai

langchain-anthropic

langchain-google-genai
google-genertiveai

langchain-huggingface
huggingface-hub

python-dotenv

numpy
scikit-learn